Finalized. Portable version added

+++ Truss_static_analysis_cpp/src/geometry_store/geom_parameters.h
@@ -16,7 +16,7 @@ class geom_parameters
 {
 public:
 	// Standard sizes
-	const float font_size = 16.0f * std::pow(10, -5);
+	const float font_size = static_cast<float>(16.0f * std::pow(10, -5));
 	const float node_circle_radii = 0.005f;
 
 	// File path
